+++ lib/k8s/client.ex
@@ -30,13 +30,18 @@ defmodule K8s.Client do
   alias K8s.Operation
   alias K8s.Client.Runner.{Async, Base, Stream, Wait}
 
+  defdelegate put_conn(operation, conn), to: Operation
+
+  @doc "alias of `K8s.Client.Runner.Base.run/1`"
+  defdelegate run(operation), to: Base
+
   @doc "alias of `K8s.Client.Runner.Base.run/2`"
   defdelegate run(conn, operation), to: Base
 
   @doc "alias of `K8s.Client.Runner.Base.run/3`"
   defdelegate run(conn, operation, http_opts), to: Base
-
   @doc "alias of `K8s.Client.Runner.Async.run/3`"
+
   defdelegate async(conn, operations), to: Async, as: :run
 
   @doc "alias of `K8s.Client.Runner.Async.run/3`"
@@ -45,17 +50,17 @@ defmodule K8s.Client do
   @doc "alias of `K8s.Client.Runner.Async.run/3`"
   defdelegate parallel(conn, operations, http_opts), to: Async, as: :run
 
+  @doc "alias of `K8s.Client.Runner.Wait.run/2`"
+  defdelegate wait_until(operation, wait_opts), to: Wait, as: :run
+
   @doc "alias of `K8s.Client.Runner.Wait.run/3`"
   defdelegate wait_until(conn, operation, wait_opts), to: Wait, as: :run
+
+  defdelegate stream(operation), to: Stream, as: :run
 
   @doc "alias of `K8s.Client.Runner.Stream.run/2`"
   defdelegate stream(conn, operation), to: Stream, as: :run
 
-  @spec stream(K8s.Conn.t(), K8s.Operation.t(), keyword) ::
-          {:error, K8s.Operation.Error.t()}
-          | {:ok,
-             ({:cont, any} | {:halt, any} | {:suspend, any}, any ->
-                :badarg | {:halted, any} | {:suspended, any, (any -> any)})}
   @doc "alias of `K8s.Client.Runner.Stream.run/3`"
   defdelegate stream(conn, operation, http_opts), to: Stream, as: :run