+++ lib/k8s/client/runner/wait.ex
@@ -34,6 +34,10 @@ defmodule K8s.Client.Runner.Wait do
   resp = K8s.Client.Runner.Wait.run(conn, op, opts)
   ```
   """
+  @spec run(Operation.t(), keyword()) ::
+          {:ok, map()} | {:error, :timeout | Error.t()}
+  def run(%Operation{conn: %Conn{} = conn} = op, opts), do: run(conn, op, opts)
+
   @spec run(Conn.t(), Operation.t(), keyword()) ::
           {:ok, map()} | {:error, :timeout | Error.t()}
   def run(%Conn{} = conn, %Operation{method: :get} = op, opts) do